backend/requirements.txt
fastapi
uvicorn
TikTokLive
redis
sqlalchemy
asyncpg
psycopg2-binary
pydantic-settings
python-dotenv